add logger library, separate launcher class and other changes to enable jar
